Fix Major performance issue
On first attempt to find tmdb_id for a library movie a cache is built of
every tmdb_id for each library entry. The recommended setting is to
persist the tmdb_id into the kodi database (for some reason not all
scrapers do this). When the tmdb_id is persisted in the kodi database,
there is a one-time hit to discover the ids, but after that it is fast.

However, that one-time-hit can be several hours! Further, it blocked all
activity on the thread that initiated the loading.
Further, the problem is worse for when the settings do not persist the
tmdbid into Kodi database, causing an awful delay on every startup of
Kodi.

Resolved issue by:
1) Discover TMDb ids on separate thread. Those requesting the info will
simply not get it with the consequence being that until the TMDb ids are
found additional information from TMDb for a movie will
not be incorporated into the user's experience. The problem will fix
itself once the TMDb data id is discovered.
2) In addition to persisting the TMDb id into the Kodi database, it is
stored in another cache- TMDbCache, which is used for other purposes
than just this one.